requirements.txt
SQLAlchemy==1.4.41
pydantic==1.10.2
fastapi==0.85.0
uvicorn==0.18.3
PyJWT==2.6.0
Pillow==9.4.0
passlib==1.7.4